% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound3.R
\name{bound3}
\alias{bound3}
\title{bound3: Bound on a uniuon of events}
\usage{
bound3(Sigma, m = NULL, m_block = 100)
}
\arguments{
\item{Sigma}{A sparse mxm corrolation matrix}

\item{m}{Dimension of the matrix Sigma}

\item{m_block}{The matrix Sigma is treated as a block diagonal
with m_block by m_block blocks. All elements outside the block are treated as zero.}
}
\value{
A function which for a vector of alpha return the bound
}
\description{
Compute a bound that the probability max_j |X_j|>PhiInv(1-alpha/2)
where X=(X_1,...,X_m)~MVN(0,Sigma)
and PhiInv is the inverse of the standard normal distribution
}
\examples{
library(Matrix)

}
\seealso{
??
}
