% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corihw.R
\name{corihw}
\alias{corihw}
\title{corihw: Main function for Independent Hypothesis Weighting with Correlation}
\usage{
corihw(pvalues, covariates, alpha, nbins = "auto", quiet = TRUE,
  nfolds = 5, lambda = 1, seed = 1L, Sigma = NULL,
  bound3_testing = TRUE, bound3_linear_testing = TRUE,
  meffective_testing = TRUE, IHW_testing = TRUE)
}
\arguments{
\item{pvalues}{Numeric vector of unadjusted p-values.}

\item{covariates}{Vector which contains the one-dimensional covariates (independent under the H0 of the p-value)
for each test. Assumed continuous.}

\item{alpha}{Numeric, sets the nominal level for FWER control.}

\item{nbins}{Integer, number of groups into which p-values will be split based on covariate. Use "auto" for
automatic selection of the number of bins. Only applicable when covariates is not a factor.}

\item{quiet}{Boolean, if False a lot of messages are printed during the fitting stages.}

\item{nfolds}{Number of folds into which the p-values will be split for the pre-validation procedure}

\item{lambda}{A regularization constant}

\item{seed}{Integer or NULL. Split of hypotheses into folds is done randomly. To have the output of the function be reproducible,
the seed of the random number generator is set to this value at the start of the function. Use NULL if you don't want to set the seed.}

\item{Sigma}{A sparse mxm corrolation matrix}

\item{bound3_testing}{Flag to perform IHW with correlation}

\item{bound3_linear_testing}{Flag to perform IHW with correlation using linear approximation.}

\item{meffective_testing}{Flag to perform IHW with correlation using m effective}

\item{IHW_testing}{Flag to perform IHW with no correlation}
}
\value{
A data frame with the weights and a rejection flag per observation
}
\description{
Given a vector of p-values, a vector of covariates which are independent of the p-values under the null hypothesis,
a matrix Sigma of the correlation of the test statistics and a nominal significance level alpha,
IHW Cor learns multiple testing weights and then applies the weighted Bonferroni) procedure.
}
