% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_example.R
\name{toy_example}
\alias{toy_example}
\title{toy_example: Creates a toy example data of multiple correlated tests}
\usage{
toy_example(m = 10000, n = 15, mu = 1, prob = 0.05,
  block_size = 100, rho = 0.8)
}
\arguments{
\item{m}{Number of tests}

\item{n}{Number of observations in each of the two samples}

\item{mu}{The mean under the alternative}

\item{prob}{The probability of alternatives}

\item{block_size}{The size of the blocks in the block-diagonal correlation matrix R.}

\item{rho}{The off-diagonal entries of th block matrices}
}
\value{
A list, with P-values (P), the pooled variance as the informative covariate (X), and the correlation matrix (Sigma).
}
\description{
Create \emph{m} two-sample t-tests with correlations. Let \emph{mu} be the mean under
the alternative and 0 the mean under the null. The proportion of alternatives is \emph{prob}.
The first sample of \emph{n}  \emph{m}-dimendional observations is drawn from a multivariate random vector with mean 0 for null
 and mean \emph{mu} for alternatives. The second sample of \emph{n} \emph{m}-dimendional observations is drawn from with mean zero.
 The correlation matrix \emph{R} is block diagonal with a given \emph{block size} and with off-diagonal entries of \emph{correlation}.
}
